Updated Cli.php::waitForKey
waitForKey requires an extra enter before the key is actually read. Since im guessing you'd like to wait for 1 key rather than 2, its been changed to passthru to just listen to a keypress event. No usages of the ouput waitForKey found in the project, so $this is returned for chaining of functions. readCharacter added to replace waitForKey's original functionality.